+++ src/IdlGenerator/UmlOperation.cpp
@@ -43,7 +43,7 @@
 static bool generate_dir(const Q3ValueList<UmlParameter> & params,
                          unsigned rank, QTextStream & f)
 {
-    if (rank >= params.count())
+    if ((int)rank >= params.count())
         return FALSE;
 
     switch (params[rank].dir) {
@@ -65,7 +65,7 @@ static bool generate_dir(const Q3ValueList<UmlParameter> & params,
 static bool generate_type(const Q3ValueList<UmlParameter> & params,
                           unsigned rank, QTextStream & f)
 {
-    if (rank >= params.count())
+    if ((int)rank >= params.count())
         return FALSE;
 
     UmlClass::write(f, params[rank].type);
@@ -75,7 +75,7 @@ static bool generate_type(const Q3ValueList<UmlParameter> & params,
 static bool generate_var(const Q3ValueList<UmlParameter> & params,
                          unsigned rank, QTextStream & f)
 {
-    if (rank >= params.count())
+    if ((int)rank >= params.count())
         return FALSE;
 
     f << params[rank].name;
